Fix #270 - Add missing CF labels for EGNOS and MSAS SBAS
